# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    description: Error response with status, code, and description
    properties:
      code:
        example: ERR_INVALID_REQUEST
        type: string
      description:
        example: Invalid request parameters
        type: string
      status:
        example: error
        type: string
    type: object
...
  handlers.ListMessagesRequest:
    description: Request to list messages from a message box
    properties:
      limit:
        description: |-
          Limit caps the number of messages returned; defaults to the server's
          configured LIST_DEFAULT_LIMIT.
        example: 100
        type: integer
      messageBox:
        example: inbox
        type: string
      messageId:
        description: |-
          MessageID, if given, restricts the page to the single message with this
          exact ID (still subject to Limit/Offset).
        example: msg-123
        type: string
      offset:
        description: Offset skips this many matching messages before the returned
          page.
        example: 0
        type: integer
      skip:
        description: Skip is a compatibility alias for Offset; if both are given they
          must match.
        example: 0
        type: integer
    type: object
  handlers.ListMessagesResponse:
    description: Response containing a page of messages plus pagination state
    properties:
      hasMore:
        description: |-
          HasMore reports whether at least one further matching message exists
          beyond this page.
        example: false
        type: boolean
      limit:
        description: |-
          Limit is the page size that was applied (the request's limit, or the
          server's configured default when none was given).
        example: 1000
        type: integer
      messages:
        items:
          $ref: '#/definitions/handlers.MessageOut'
        type: array
      nextOffset:
        description: |-
          NextOffset is Offset plus the number of messages actually returned; pass
          it as the next request's offset to continue paging.
        example: 1000
        type: integer
      offset:
        description: Offset is the request's offset (or skip), echoed back.
        example: 0
        type: integer
      status:
        example: success
        type: string
    type: object
  handlers.ListPermissionsResponse:
    description: Response containing list of permissions
    properties:
      permissions:
        items:
          $ref: '#/definitions/handlers.PermissionDetailList'
        type: array
      status:
        example: success
        type: string
      totalCount:
        example: 10
        type: integer
    type: object
  handlers.MessageOut:
    description: Message object returned by listMessages
    properties:
      body:
        example: '{"text":"Hello"}'
        type: string
      createdAt:
        example: "2024-01-01T12:00:00.000Z"
        type: string
      messageId:
        example: abc123
        type: string
      sender:
        example: 03abc...
        type: string
      updatedAt:
        example: "2024-01-01T12:00:00.000Z"
        type: string
    type: object
...
  handlers.PermissionDetailList:
    description: Permission details (snake_case for listPermissions endpoint)
    properties:
      created_at:
        example: "2024-01-01T12:00:00.000Z"
        type: string
      message_box:
        example: inbox
        type: string
      recipient_fee:
        example: 100
        type: integer
      sender:
        example: 03abc...
        type: string
      updated_at:
        example: "2024-01-01T12:00:00.000Z"
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for message delivery, retrieval, acknowledgement and permissions.
    Uses BRC-31/BRC-104 mutual authentication.
  title: MessageBox Server API
  version: 1.0.0
paths:
...
  /api/handle:
    put:
      consumes:
      - application/json
      parameters:
      - description: Self-signed profile certificate
        in: body
        name: certificate
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Register, update or release a handle
      tags:
      - lookup
  /api/handle/{query}:
    get:
      parameters:
      - description: Partial or full handle, optionally @domain
        in: path
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Search handles (capability 0ace65da5987)
      tags:
      - lookup
...
  /listMessages:
    post:
      consumes:
      - application/json
      description: Returns one deterministic, bounded page of stored messages for
        the specified messageBox belonging to the authenticated identity, most recently
        paginated with limit/offset (skip is a compatibility alias for offset) and
        optionally filtered to a single messageId. If the box does not exist or has
        no messages, an empty page is returned. Existing clients that send only messageBox
        get the first page of up to the server's default limit.
      parameters:
      - description: Message box to list messages from, with optional pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ListMessagesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListMessagesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "413":
          description: the oldest message in the page exceeds the configured response
            byte budget
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BSVAuth: []
      summary: Retrieve messages from a message box
      tags:
      - Messages
...
  /permissions/list:
    get:
      description: Returns all permission settings for the authenticated identity,
        optionally filtered by message box.
      parameters:
      - description: Filter by message box name
        in: query
        name: messageBox
        type: string
      - description: Maximum number of results (1-1000, default 100)
        in: query
        name: limit
        type: integer
      - description: Number of results to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: 'Sort order: ''asc'' or ''desc'' (default ''desc'')'
        in: query
        name: createdAtOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListPermissionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BSVAuth: []
      summary: List message permissions
      tags:
      - Permissions
...
securityDefinitions:
  BSVAuth:
    description: BRC-31/BRC-104 mutual authentication. Requires multiple x-bsv-auth-*
      headers (identity-key, nonce, signature, etc.)
    in: header
    name: x-bsv-auth-identity-key
    type: apiKey
swagger: "2.0"